requirements.txt
altair==4.2.2
gensim==4.1.2
gensim==4.3.0
google_trans_new==1.1.9
gTTS==2.3.2
langchain==0.0.161
matplotlib==3.5.1
matplotlib==3.6.2
matplotlib==3.7.0
matplotlib==3.6.3
openai==0.27.2
openpyxl==3.0.9
openpyxl==3.0.10
pandas==1.3.4
pandas==1.4.2
pandas==1.4.4
pandas==1.5.3
pandas_profiling==3.6.6
python-dotenv==1.0.0
spacy==3.4.4
spacy_streamlit==1.0.6
streamlit==1.22.0
streamlit_chat==0.0.2.2
streamlit_option_menu==0.3.2
streamlit_pandas_profiling==0.1.3
textblob==0.17.1
vaderSentiment==3.3.2
vaderSentiment==3.3.2
wordcloud==1.9.1.1
ydata_profiling==4.1.2
numpy==1.23.5

